requirements.txt


altgraph==0.17.4
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dj-database-url==3.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-heroku==0.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
channels==4.1.0
daphne==4.1.2
websockets==12.0
filelock==3.18.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.4.0
MarkupSafe==3.0.2
nameparser==1.1.3
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyJWT==2.9.0
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.1
pywin32-ctypes==0.2.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.1
virtualenv==20.31.2
Werkzeug==3.1.3
whitenoise==6.6.0
nltk>=3.9



